refactor: added it to hooks
